feat: profile display by modules has been modified
